Update uv.lock using `uv sync` (#23)
Updates lockfile to the latest uv format. In 0.4.27, uv started
supporting the dependency-groups package metadata specifier
(https://raw.githubusercontent.com/astral-sh/uv/main/CHANGELOG.md), so
this updates to that format.

There are some changes, such as resolving websockets to >=10.4, that
make it more closely track the dependencies that we have specified in
pyproject.toml.

In any case, this file is transparent to us.